requirements.txt
annotated-types==0.7.0
anyio==4.6.2.post1
gitignore_parser==0.1.11
idna==3.10
importlib_metadata==8.5.0
nats-py==2.9.0
packaging==24.1
pycron==3.1.1
pydantic==2.9.2
pydantic_core==2.23.4
pytz==2024.2
sniffio==1.3.1
taskiq==0.11.7
taskiq-dependencies==1.5.4
taskiq-nats==0.4.0
typing_extensions==4.12.2
watchdog==2.3.1
zipp==3.20.2
